+++ ft-sdk/src/from_request/form.rs
@@ -5,3 +5,9 @@ impl<T: serde::de::DeserializeOwned> ft_sdk::FromRequest for Form<T> {
         Ok(serde_json::from_value(req.body().clone()).map(Form)?)
     }
 }
+
+impl<T: serde::de::DeserializeOwned> AsRef<T> for Form<T> {
+    fn as_ref(&self) -> &T {
+        &self.0
+    }
+}