+++ ft-sdk/src/from_request/scheme.rs
@@ -1,3 +1,4 @@
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum Scheme {
     Http,
     Https,
@@ -14,10 +15,31 @@ impl std::fmt::Display for Scheme {
 
 impl ft_sdk::FromRequest for Scheme {
     fn from_request(req: &http::Request<serde_json::Value>) -> ft_sdk::Result<Scheme> {
-        if req.uri().scheme_str().unwrap_or_default() == "https" {
+        let scheme = req.uri().scheme_str().unwrap_or_default();
+        if scheme == "wasm+proxy" {
+            // This is a bug fix for fastn, in fastn we pass the scheme as wasm+proxy, it should
+            // be http or https. This is a temporary fix until we fix fastn.
+            let host: ft_sdk::Host = ft_sdk::Host::from_request(req)?;
+            return if host.without_port() == "127.0.0.1" {
+                Ok(Scheme::Http)
+            } else {
+                Ok(Scheme::Https)
+            };
+        }
+
+        if scheme == "https" {
             return Ok(Scheme::Https);
         }
 
         Ok(Scheme::Http)
+    }
+}
+
+impl AsRef<str> for Scheme {
+    fn as_ref(&self) -> &str {
+        match self {
+            Scheme::Http => "http",
+            Scheme::Https => "https",
+        }
     }
 }